feat: Merge analytics configs
