--- src/Firebase/firebase-config.ts
@@ -1,12 +1,0 @@
-// Firebase configuration for analytics
-// Used by @curiouslearning/analytics FirebaseStrategy
-export const firebaseConfig = {
-    apiKey: "AIzaSyB8c2lBVi26u7YRL9sxOP97Uaq3yN8hTl4",
-    authDomain: "ftm-b9d99.firebaseapp.com",
-    databaseURL: "https://ftm-b9d99.firebaseio.com",
-    projectId: "ftm-b9d99",
-    storageBucket: "ftm-b9d99.appspot.com",
-    messagingSenderId: "602402387941",
-    appId: "1:602402387941:web:a63f4eaddc949f539de10c",
-    measurementId: "G-FVLSN7D7NM",
-};
