--- src/Firebase/statsig-config.ts
@@ -1,11 +1,0 @@
-// Statsig configuration for analytics
-// Used by @curiouslearning/analytics StatsigStrategy
-// TODO: Replace with actual Statsig configuration values
-export const statsigConfig = {
-    // Statsig client-side API key
-    clientKey: "client-SSmY5k5Cs39G7II74NdWqPfv5hQzrFiUqCc3C1IU9na",
-    
-    // User ID for Statsig tracking (can be pseudo-anonymous)
-    userId: "anonymous-user",
-    
-};
